+++ market_gamble_detector.py
@@ -13,12 +13,6 @@
 
 import time
 
-from drop_rate_estimates import (
-    clamp_proportion,
-    get_drop_rate_estimates_based_on_item_rarity_pattern,
-    get_drop_rate_field,
-    get_rarity_fields,
-)
 from market_arbitrage import (
     filter_out_badges_with_low_sell_price,
     find_badge_arbitrages,
@@ -29,12 +23,18 @@ from market_buzz_detector import (
     print_packs_with_high_buzz,
     sort_according_to_buzz,
 )
-from market_listing import get_item_nameid_batch
-from market_order import (
+from src.drop_rate_estimates import (
+    clamp_proportion,
+    get_drop_rate_estimates_based_on_item_rarity_pattern,
+    get_drop_rate_field,
+    get_rarity_fields,
+)
+from src.market_listing import get_item_nameid_batch
+from src.market_order import (
     download_market_order_data_batch,
     load_market_order_data_from_disk,
 )
-from market_search import (
+from src.market_search import (
     get_steam_api_rate_limits_for_market_search,
     get_tag_item_class_no_for_emoticons,
     get_tag_item_class_no_for_profile_backgrounds,
@@ -42,9 +42,9 @@ from market_search import (
     load_all_listings,
     update_all_listings,
 )
-from personal_info import get_cookie_dict
-from sack_of_gems import get_gem_amount_required_to_craft_badge, get_gem_price
-from utils import (
+from src.personal_info import get_cookie_dict
+from src.sack_of_gems import get_gem_amount_required_to_craft_badge, get_gem_price
+from src.utils import (
     convert_listing_hash_to_app_id,
     get_category_name_for_booster_packs,
     get_category_name_for_emoticons,
