+++ src/creation_time_utils.py
@@ -1,7 +1,7 @@
 import datetime
 
 from src.json_utils import load_json
-from utils import get_next_creation_time_file_name
+from src.utils import get_next_creation_time_file_name
 
 
 def load_next_creation_time_data(
